+++ server/api/v1/client/phone_area_code.go
@@ -6,6 +6,7 @@ import (
 	clientReq "github.com/flipped-aurora/gin-vue-admin/server/model/client/request"
 	"github.com/flipped-aurora/gin-vue-admin/server/model/common/response"
 	"github.com/flipped-aurora/gin-vue-admin/server/service"
+	"github.com/flipped-aurora/gin-vue-admin/server/utils"
 	"github.com/flipped-aurora/gin-vue-admin/server/utils/i18n"
 	"github.com/gin-gonic/gin"
 	"go.uber.org/zap"
@@ -25,6 +26,10 @@ var phoneAreaCodeService = service.ServiceGroupApp.ClientServiceGroup.PhoneAreaC
 // @Success 200 {string} string "{"success":true,"data":{},"msg":"创建成功"}"
 // @Router /phoneAreaCode/createPhoneAreaCode [post]
 func (api *PhoneAreaCodeApi) CreatePhoneAreaCode(c *gin.Context) {
+	if !isClientAdminAuthority(utils.GetUserAuthorityId(c)) {
+		failClientWithKey(c, "noPermission")
+		return
+	}
 	var info client.PhoneAreaCode
 	err := c.ShouldBindJSON(&info)
 	if err != nil {
@@ -49,6 +54,10 @@ func (api *PhoneAreaCodeApi) CreatePhoneAreaCode(c *gin.Context) {
 // @Success 200 {string} string "{"success":true,"data":{},"msg":"删除成功"}"
 // @Router /phoneAreaCode/deletePhoneAreaCode [delete]
 func (api *PhoneAreaCodeApi) DeletePhoneAreaCode(c *gin.Context) {
+	if !isClientAdminAuthority(utils.GetUserAuthorityId(c)) {
+		failClientWithKey(c, "noPermission")
+		return
+	}
 	ID := c.Query("ID")
 	if err := phoneAreaCodeService.DeletePhoneAreaCode(ID); err != nil {
 		global.GVA_LOG.Error("删除失败!", zap.Error(err))
@@ -68,6 +77,10 @@ func (api *PhoneAreaCodeApi) DeletePhoneAreaCode(c *gin.Context) {
 // @Success 200 {string} string "{"success":true,"data":{},"msg":"更新成功"}"
 // @Router /phoneAreaCode/updatePhoneAreaCode [put]
 func (api *PhoneAreaCodeApi) UpdatePhoneAreaCode(c *gin.Context) {
+	if !isClientAdminAuthority(utils.GetUserAuthorityId(c)) {
+		failClientWithKey(c, "noPermission")
+		return
+	}
 	var info client.PhoneAreaCode
 	err := c.ShouldBindJSON(&info)
 	if err != nil {
@@ -92,6 +105,10 @@ func (api *PhoneAreaCodeApi) UpdatePhoneAreaCode(c *gin.Context) {
 // @Success 200 {object} response.Response{data=response.PageResult,msg=string} "获取成功"
 // @Router /phoneAreaCode/getPhoneAreaCodeList [get]
 func (api *PhoneAreaCodeApi) GetPhoneAreaCodeList(c *gin.Context) {
+	if !isClientAdminAuthority(utils.GetUserAuthorityId(c)) {
+		failClientWithKey(c, "noPermission")
+		return
+	}
 	var pageInfo clientReq.PhoneAreaCodeSearch
 	err := c.ShouldBindQuery(&pageInfo)
 	if err != nil {
